+++ app/(main)/(index)/Intro.tsx
