+++ app/(main)/(index)/LetsJoin.tsx
