+++ app/(main)/(index)/ScrollDownAnimation.tsx
