+++ app/(main)/(index)/Slide.tsx
