+++ app/(main)/Banner.tsx
