+++ app/(main)/guide/Intro.tsx
