+++ app/(main)/guide/Miner.tsx
