+++ app/(main)/menu/icons/Newbie.tsx
