+++ app/(main)/menu/icons/Psi.tsx
