+++ app/(main)/newbie/HowToParticipate.tsx
