+++ app/(main)/newbie/Intro.tsx
