+++ app/(main)/newbie/Rule.tsx
