+++ app/(main)/newbie/WhatToDo.tsx
