+++ app/(main)/newbie/WhenTrouble.tsx
