+++ app/(main)/news/Banner.tsx
